requirements.txt
aiohttp==3.7.2
aitoai==0.4.0
argcomplete==1.12.1
asgiref==3.3.0
async-timeout==3.0.1
attrs==20.3.0
boto3==1.16.12
botocore==1.19.12
certifi==2020.6.20
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.5.0
django-extensions==3.0.9
django-import-export==2.4.0
django-storages==1.9.1
djangorestframework==3.12.2
docutils==0.16
drf-yasg==1.20.0
durationpy==0.5
et-xmlfile==1.0.1
future==0.18.2
hjson==3.0.2
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
kappa==0.6.0
lambda-packages==0.20.0
langdetect==1.0.8
MarkupPy==1.14
MarkupSafe==1.1.1
multidict==5.0.0
ndjson==0.3.1
numpy==1.19.4
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.4
pandas==1.1.4
pip-tools==5.3.1
placebo==0.9.0
pydash==4.9.0
PyMySQL==0.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.51.0
troposphere==2.6.3
typing-extensions==3.7.4.3
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
xlrd==1.2.0
xlwt==1.3.0
yarl==1.6.2
zappa==0.48.2
zipp==3.4.0
